Refactor panel -> loader project
