+++ api-model/src/main/resources/application.conf
@@ -1,0 +1,3 @@
+hmda {
+  connectionFlowParallelism = 5
+}
