+++ loader/src/main/resources/application.conf
@@ -7,5 +7,6 @@ akka {
 hmda {
   adminUrl = "http://0.0.0.0:8081"
   adminUrl = ${?HMDA_HTTP_ADMIN_URL}
+  httpTimeout = 5
 }
